pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nonet-movie"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
    "underpyx",
    "pydmpro",
    "dddx",
    "rich",
    "questionary",
    "prompt-toolkit",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "assertpy"
]

[tool.setuptools]
package-dir = {"" = "src"}